requirements.txt
langchain==0.3.26
flask==3.1.1
flask-cors==4.0.1
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-google-genai==2.1.12
langchain-community==0.3.26
pinecone
gunicorn==22.0.0
requests
-e .